src/vision/__init__.py
"""
Vision Module

Handles screen capture, image recognition, and game state detection.
Uses computer vision techniques to understand the current state of the game.
"""

import cv2
import numpy as np
from PIL import ImageGrab


class ScreenCapture:
    """Captures screenshots of the game window."""
    
    def __init__(self):
        self.game_window = None
    
    def capture(self):
        """Capture the current game screen."""
        # TODO: Implement screen capture
        pass


class ImageRecognition:
    """Recognizes game elements in screenshots."""
    
    def __init__(self):
        self.templates = {}
    
    def load_templates(self, template_dir):
        """Load reference images for template matching."""
        # TODO: Implement template loading
        pass
    
    def detect_resources(self, image):
        """Detect resource levels (food, water, power) from image."""
        # TODO: Implement resource detection
        pass
    
    def detect_dwellers(self, image):
        """Detect dweller positions and states."""
        # TODO: Implement dweller detection
        pass
